LC-valid-perfect-square [Runtime: N/A; Memory: N/A]

+++ math/valid-perfect-square.cpp
@@ -4,7 +4,7 @@ public:
         int low = 0;
         int high = num/2+1;
         while(low <= high){
-            int mid = low +(high -low )/2;
+            long long mid = low +(high -low )/2;
             if(mid * mid == num) return true;
             else if (mid*mid > num) high = mid - 1;
             else low = mid + 1;
